--- methods/mod.ts
@@ -1,3 +1,0 @@
-export * from "./create_element.ts";
-export * from "./to_text.ts";
-export * from "./write_file.ts";
